# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "panmira N1 API",
    "version": "1.6.0",
    "description": "panmira N1 OpenAPI 3.0 spec. IA v6 (commit 95b16907 + IA v6 alignment, 2026-07-08). New IA v6 module paths under /api/v2/{overview,people,employees,tasks,foundation,channels,models}. Old paths under /api/v2/admin/* carry Deprecation: true + Sunset: 2026-08-01 headers.",
    "contact": {
      "name": "panmira"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9100",
      "description": "Local dev"
    },
    {
      "url": "https://deepx.fun",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/v2/admin/channels": {
      "get": {
        "summary": "List channels",
        "description": "[DEPRECATED 2026-07-08 — Sunset 2026-08-01] Use new IA v6 path. List configured channels. Scope: channel:read.",
        "tags": [
          "admin-channels"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "deprecated": true
      },
      "post": {
        "summary": "Create channel",
        "description": "[DEPRECATED 2026-07-08 — Sunset 2026-08-01] Use new IA v6 path. Register a new channel. Scope: channel:admin.",
        "tags": [
          "admin-channels"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "deprecated": true
      }
    },
...
    "/api/v2/channels": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "资源频道·接入点列表",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "feishu",
                "telegram",
                "wechat",
                "webhook",
                "cli",
                "custom"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth 2.0 Bearer token (access_token)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {},
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
...
  }
}